+++ demo/src/main/java/com/novo/main/SampleChooserActivity.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.android.exoplayer2.demo;
+package com.novo.main;
 
 import android.app.Activity;
 import android.content.Context;
@@ -40,6 +40,8 @@ import com.google.android.exoplayer2.upstream.DataSpec;
 import com.google.android.exoplayer2.upstream.DefaultDataSource;
 import com.google.android.exoplayer2.util.Assertions;
 import com.google.android.exoplayer2.util.Util;
+import com.novo.R;
+
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;