+++ demo/src/main/java/com/novo/network/EndPoints.java
@@ -1,0 +1,13 @@
+package com.novo.network;
+
+/**
+ * Created by Hisham on 08/Sep/2017 - 14:17
+ */
+
+public class EndPoints {
+//    private static final String baseUrl = "http://35.154.11.202/";
+    private static final String baseUrl = "http://192.168.0.94:8080/";
+    public static String getBaseUrl() {
+        return baseUrl;
+    }
+}