+++ demo/src/main/java/com/novo/network/ServerHit.java
@@ -1,8 +1,12 @@
-package com.google.android.exoplayer2.demo;
+package com.novo.network;
 
+import android.app.Activity;
+import android.app.ProgressDialog;
 import android.os.AsyncTask;
 import android.text.TextUtils;
 import android.util.Log;
+
+import com.google.android.exoplayer2.upstream.novo.TokenManager;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -11,6 +15,8 @@ import java.io.InputStreamReader;
 import java.io.OutputStream;
 import java.net.HttpURLConnection;
 import java.net.URL;
+
+import static com.novo.util.Utils.TAG;
 
 /**
  * Created by Hisham on 25/Aug/2017 - 20:01
@@ -30,17 +36,29 @@ public class ServerHit {
         private final String type;
         private String body;
         private String contentType;
+        private final Activity activity;
+        private ProgressDialog progressDialog;
+        private String token;
 
-        public JSONTask(String type, String contentType, String body, ServiceHitResponseListener listener){
+        public JSONTask(Activity activity, String token, String type, String contentType, String body, ServiceHitResponseListener listener){
             this.body = body;
             this.listener = listener;
             this.type = type;
             this.contentType = contentType;
+            this.activity = activity;
+            this.token = token;
         }
 
         @Override
         protected void onPreExecute() {
             super.onPreExecute();
+            // hit server to get all videos urls
+            progressDialog = new ProgressDialog(activity);
+            progressDialog.setMessage("Please wait...");
+            progressDialog.setIndeterminate(true);
+            progressDialog.setProgressStyle(ProgressDialog.STYLE_SPINNER);
+            progressDialog.setCancelable(false);
+            progressDialog.show();
         }
 
         @Override
@@ -56,6 +74,9 @@ public class ServerHit {
                 }
                 if(!TextUtils.isEmpty(contentType)) {
                     connection.setRequestProperty("Content-Type", contentType);
+                }
+                if(!TextUtils.isEmpty(token)) {
+                    connection.setRequestProperty("token", TokenManager.getToken());
                 }
                 if(body != null) {
                     byte[] outputInBytes = body.getBytes("UTF-8");
@@ -103,11 +124,7 @@ public class ServerHit {
                 listener.onError(null);
 //                Toast.makeText(getApplicationContext(), "Not able to fetch data from server, please check url.", Toast.LENGTH_SHORT).show();
             }
+            progressDialog.dismiss();
         }
     }
-
-    private static final String TAG = ServerHit.class.getSimpleName();
-
-
-    
 }